% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rg_summary.R
\name{print.rg_summary}
\alias{print.rg_summary}
\title{Print rg_summary objects}
\usage{
\method{print}{rg_summary}(x, ...)
}
\arguments{
\item{x}{An rg_summary object.}

\item{...}{Unused.}
}
\value{
Invisibly returns x.
}
\description{
Print rg_summary objects
}
