% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rgr_summary.R
\name{print.rgr_summary}
\alias{print.rgr_summary}
\title{Print rgr_summary objects}
\usage{
\method{print}{rgr_summary}(x, ...)
}
\arguments{
\item{x}{An rgr_summary object.}

\item{...}{Unused.}
}
\value{
Invisibly returns x.
}
\description{
Prints a compact quality-assurance and descriptive summary produced by
\code{rgr_summary()}.
}
