% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_add_risk_attributes.R
\name{rg_add_risk_attributes}
\alias{rg_add_risk_attributes}
\title{Add synthetic risk attributes to spatial features}
\usage{
rg_add_risk_attributes(x, target = NULL, mode = NULL, seed = NULL)
}
\arguments{
\item{x}{An sf object.}

\item{target}{Character. One of \code{"points"} or \code{"polygons"}. If NULL, inferred from geometry.}

\item{mode}{Character. One of \code{"case"} or \code{"count"}. If NULL, inferred from geometry.}

\item{seed}{Optional random seed.}
}
\value{
An sf object with synthetic risk attributes.
}
\description{
Add synthetic risk attributes to spatial features
}
