% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_example_data.R
\name{rg_example_data}
\alias{rg_example_data}
\title{Return built-in example datasets and scenarios}
\usage{
rg_example_data(name = "wa_outback")
}
\arguments{
\item{name}{Name of the example dataset or scenario.}
}
\value{
An sf object or list depending on the example selected.
}
\description{
Return built-in example datasets and scenarios
}
