% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_grid.R
\name{rg_grid}
\alias{rg_grid}
\title{Generate a square grid over a study area}
\usage{
rg_grid(study_area, cell_size = 500, clip = FALSE)
}
\arguments{
\item{study_area}{An sf polygon object.}

\item{cell_size}{Grid cell size in map units.}

\item{clip}{Logical. Should the grid be clipped to the study area? Defaults to FALSE.}
}
\value{
An sf polygon grid.
}
\description{
Generate a square grid over a study area
}
