% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_hex.R
\name{rg_hex}
\alias{rg_hex}
\title{Generate a hexagon grid over a study area}
\usage{
rg_hex(study_area, cell_size = 500, clip = FALSE)
}
\arguments{
\item{study_area}{An sf polygon object.}

\item{cell_size}{Approximate hexagon cell size in map units.}

\item{clip}{Logical. Should the hex grid be clipped to the study area? Defaults to FALSE.}
}
\value{
An sf polygon hex grid.
}
\description{
Generate a hexagon grid over a study area
}
