% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgr_bbox_example.R
\name{rgr_bbox_example}
\alias{rgr_bbox_example}
\title{Return example bounding boxes for riskGenerateR}
\usage{
rgr_bbox_example(name = "generic")
}
\arguments{
\item{name}{Example name. One of \code{"generic"}, \code{"wa_outback"}, or \code{"vicgrid"}.}
}
\value{
A list containing bbox, crs, name, and description.
}
\description{
Return example bounding boxes for riskGenerateR
}
