% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgr_topology_modify.R
\name{rgr_topology_modify}
\alias{rgr_topology_modify}
\title{Modify polygon topology for QA and stress testing}
\usage{
rgr_topology_modify(x, mode = "overlap", pct = 0.1, distance = 25, seed = NULL)
}
\arguments{
\item{x}{An sf polygon object.}

\item{mode}{Topology modification mode. Currently only \code{"overlap"}.}

\item{pct}{Proportion of features to modify.}

\item{distance}{Buffer distance in map units.}

\item{seed}{Optional random seed.}
}
\value{
An sf object with intentionally modified topology.
}
\description{
Modify polygon topology for QA and stress testing
}
