% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_points.R
\name{rg_points}
\alias{rg_points}
\title{Generate random points inside and outside a study area}
\usage{
rg_points(
  study_area,
  n = 100,
  inside_pct = 0,
  outside_distance = 1000,
  seed = NULL,
  processing_crs = NULL,
  return_input_crs = TRUE
)
}
\arguments{
\item{study_area}{An sf polygon object.}

\item{n}{Number of points to generate.}

\item{inside_pct}{Proportion of points generated inside the study area. Defaults to 0.}

\item{outside_distance}{Distance in map units of the processing CRS.}

\item{seed}{Optional random seed.}

\item{processing_crs}{Optional projected CRS used for point generation.}

\item{return_input_crs}{Logical. Return output to the input CRS? Defaults to TRUE.}
}
\value{
An sf point object.
}
\description{
Generate random points inside and outside a study area
}
