% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_polygons.R
\name{rg_polygons}
\alias{rg_polygons}
\title{Generate irregular synthetic polygon units}
\usage{
rg_polygons(
  study_area,
  target_n = 20,
  cell_size = 250,
  method = "seeded_grid",
  clip = TRUE,
  seed = NULL,
  processing_crs = NULL,
  return_input_crs = TRUE
)
}
\arguments{
\item{study_area}{An sf polygon object.}

\item{target_n}{Target number of polygon units.}

\item{cell_size}{Fine grid cell size in map units of the processing CRS.}

\item{method}{Polygon generation method. Currently only \code{"seeded_grid"}.}

\item{clip}{Logical. Should output polygons be clipped to the study area?}

\item{seed}{Optional random seed.}

\item{processing_crs}{Optional projected CRS used for geometry generation.}

\item{return_input_crs}{Logical. Return output to the input CRS? Defaults to TRUE.}
}
\value{
An sf polygon object.
}
\description{
Generate irregular synthetic polygon units
}
