% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_summary.R
\name{rg_summary}
\alias{rg_summary}
\title{Summarise a generated riskGenerateR spatial object}
\usage{
rg_summary(x)
}
\arguments{
\item{x}{An sf object.}
}
\value{
A named list containing summary values.
}
\description{
Summarise a generated riskGenerateR spatial object
}
